+++ src/pages/onboarding/Alarm.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useNavigate } from 'react-router';
 import ProgressHeader from '@/components/ui/progressheader';
-import Button from '@/componenets/Button/Button';
+import { Button } from '@/components/ui/button';
 
 export const Alarm: React.FC = () => {
   const navigate = useNavigate();
