+++ src/pages/onboarding/Birthday.tsx
@@ -1,15 +1,10 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router';
 import ProgressHeader from '@/components/ui/progressheader';
-import Button from '@/componenets/Button/Button';
+import { Button } from '@/components/ui/button';
 import Input from '@/componenets/Input/Input';
-import Picker from 'react-mobile-picker';
 import { Checkbox } from '@/components/ui/checkbox';
-import {
-  Drawer,
-  DrawerContent,
-  DrawerHeader,
-} from '@/components/ui/drawer';
+import { DatePickerDrawer } from '@/components/ui/datepicker';
 
 export const Birthday: React.FC = () => {
   const [birthday, setBirthday] = useState('');
@@ -18,28 +13,17 @@ export const Birthday: React.FC = () => {
   const [isGenderVisible, setIsGenderVisible] = useState(false);
   const navigate = useNavigate();
 
-  const today = new Date();
-  const years = Array.from({ length: 100 }, (_, i) => String(today.getFullYear() - i));
-  const months = Array.from({ length: 12 }, (_, i) => String(i + 1));
-  const days = Array.from({ length: 31 }, (_, i) => String(i + 1));
-
-
-  const [pickerValue, setPickerValue] = useState({
-    year: String(today.getFullYear()),
-    month: String(today.getMonth() + 1),
-    day: String(today.getDate()),
-  });
-
   const handleSubmit = () => {
     if (birthday) {
       navigate('/onboarding/alarm');
     }
   };
 
-  const handleConfirm = () => {
-    const { year, month, day } = pickerValue;
+  const handleDateConfirm = (date: Date) => {
+    const year = date.getFullYear();
+    const month = String(date.getMonth() + 1).padStart(2, '0');
+    const day = String(date.getDate()).padStart(2, '0');
     setBirthday(`${year}년 ${month}월 ${day}일`);
-    setIsDrawerOpen(false);
   };
 
   return (
@@ -83,7 +67,6 @@ export const Birthday: React.FC = () => {
       </div>
 
       <div className="p-4">
-
         <Button
           text="다음"
           variant={birthday && (isAgeVisible && isGenderVisible) ? 'active' : 'inactive'}
@@ -92,57 +75,11 @@ export const Birthday: React.FC = () => {
         />
       </div>
 
-      <Drawer open={isDrawerOpen} onOpenChange={setIsDrawerOpen}>
-        <DrawerContent>
-          <div className="flex flex-col items-center w-full">
-            <div className="w-full h-[200px] flex items-center justify-center">
-              <Picker value={pickerValue} onChange={setPickerValue} height={180} itemHeight={44}>
-                <Picker.Column name="year">
-                  {years.map(year => (
-                    <Picker.Item key={year} value={year}>
-                      {({ selected }) => (
-                        <div className={`w-24 text-center ${selected ? 'text-black font-semibold text-lg' : 'text-gray-400'}`}>
-                          {year}
-                        </div>
-                      )}
-                    </Picker.Item>
-                  ))}
-                </Picker.Column>
-                <Picker.Column name="month">
-                  {months.map(month => (
-                    <Picker.Item key={month} value={month}>
-                      {({ selected }) => (
-                        <div className={`w-24 text-center ${selected ? 'text-black font-semibold text-lg' : 'text-gray-400'}`}>
-                          {month}
-                        </div>
-                      )}
-                    </Picker.Item>
-                  ))}
-                </Picker.Column>
-                <Picker.Column name="day">
-                  {days.map(day => (
-                    <Picker.Item key={day} value={day}>
-                      {({ selected }) => (
-                        <div className={`w-24 text-center ${selected ? 'text-black font-semibold text-lg' : 'text-gray-400'}`}>
-                          {day}
-                        </div>
-                      )}
-                    </Picker.Item>
-                  ))}
-                </Picker.Column>
-              </Picker>
-            </div>
-          </div>
-          <DrawerHeader >
-            <Button
-              text="확인"
-              variant="active"
-              onClick={handleConfirm}
-              className="w-full"
-            />
-          </DrawerHeader>
-        </DrawerContent>
-      </Drawer>
+      <DatePickerDrawer
+        isOpen={isDrawerOpen}
+        onOpenChange={setIsDrawerOpen}
+        onConfirm={handleDateConfirm}
+      />
     </div>
   );
 };
