+++ src/pages/onboarding/FirstMeet.tsx
@@ -1,32 +1,13 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router';
 import ProgressHeader from '../../components/ui/progressheader';
-import Button from '@/componenets/Button/Button';
+import { Button } from '@/components/ui/button';
 import Input from '@/componenets/Input/Input';
-import Picker from 'react-mobile-picker';
-import {
-  Drawer,
-  DrawerContent,
-  DrawerHeader,
-} from '@/components/ui/drawer'; // shadcn drawer
+import { DatePickerDrawer } from '@/components/ui/datepicker';
 
 export const FirstMeet: React.FC = () => {
   const navigate = useNavigate();
   const [isDrawerOpen, setIsDrawerOpen] = useState(false);
-
-  const today = new Date();
-const years = Array.from({ length: 100 }, (_, i) => String(today.getFullYear() - i));
-const months = Array.from({ length: 12 }, (_, i) => String(i + 1));
-const days = Array.from({ length: 31 }, (_, i) => String(i + 1));
-
-
-const [pickerValue, setPickerValue] = useState({
-  year: String(today.getFullYear()),
-  month: String(today.getMonth() + 1),
-  day: String(today.getDate()),
-});
-
-
   const [selectedDate, setSelectedDate] = useState<Date | null>(null);
 
   const formatDate = (date: Date) => {
@@ -42,11 +23,8 @@ const [pickerValue, setPickerValue] = useState({
     }
   };
 
-  const handleConfirm = () => {
-    const { year, month, day } = pickerValue;
-    const date = new Date(Number(year), Number(month) - 1, Number(day));
+  const handleDateConfirm = (date: Date) => {
     setSelectedDate(date);
-    setIsDrawerOpen(false);
   };
 
   return (
@@ -67,11 +45,6 @@ const [pickerValue, setPickerValue] = useState({
           status={selectedDate ? 'active' : 'inactive'}
           onClear={() => {
             setSelectedDate(null);
-            setPickerValue({
-              year: String(today.getFullYear()),
-              month: String(today.getMonth() + 1),
-              day: String(today.getDate()),
-            });
           }}
           className="w-full border border-gray-200 rounded-lg px-4 py-3 text-left text-gray-900 placeholder-gray-400 text-base focus:outline-none focus:ring-2 focus:ring-green-500"
         />
@@ -86,58 +59,12 @@ const [pickerValue, setPickerValue] = useState({
         />
       </div>
 
-      <Drawer open={isDrawerOpen} onOpenChange={setIsDrawerOpen}>
-        <DrawerContent >
-          
-          <div className="flex flex-col items-center w-full">
-            <div className="w-full  h-[200px] flex items-center justify-center">
-              <Picker value={pickerValue} onChange={setPickerValue} height={180} itemHeight={44}>
-                <Picker.Column name="year">
-                  {years.map(year => (
-                    <Picker.Item key={year} value={year}>
-                      {({ selected }) => (
-                        <div className={`w-24 text-center ${selected ? 'text-black font-semibold text-lg' : 'text-gray-400'}`}>
-                          {year}
-                        </div>
-                      )}
-                    </Picker.Item>
-                  ))}
-                </Picker.Column>
-                <Picker.Column name="month">
-                  {months.map(month => (
-                    <Picker.Item key={month} value={month}>
-                      {({ selected }) => (
-                        <div className={`w-24 text-center ${selected ? 'text-black font-semibold text-lg' : 'text-gray-400'}`}>
-                          {month}
-                        </div>
-                      )}
-                    </Picker.Item>
-                  ))}
-                </Picker.Column>
-                <Picker.Column name="day">
-                  {days.map(day => (
-                    <Picker.Item key={day} value={day}>
-                      {({ selected }) => (
-                        <div className={`w-24 text-center ${selected ? 'text-black font-semibold text-lg' : 'text-gray-400'}`}>
-                          {day}
-                        </div>
-                      )}
-                    </Picker.Item>
-                  ))}
-                </Picker.Column>
-              </Picker>
-            </div>
-          </div><DrawerHeader className="p-4">
-            <Button
-              text="확인"
-              variant="active"
-              onClick={handleConfirm}
-              className="w-full"
-            />
-          </DrawerHeader>
-        </DrawerContent>
-        
-      </Drawer>
+      <DatePickerDrawer
+        isOpen={isDrawerOpen}
+        onOpenChange={setIsDrawerOpen}
+        onConfirm={handleDateConfirm}
+        initialDate={selectedDate || undefined}
+      />
     </div>
   );
 };
