+++ src/pages/onboarding/MilitaryDay.tsx
@@ -1,14 +1,9 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router';
 import ProgressHeader from '../../components/ui/progressheader';
-import Button from '@/componenets/Button/Button';
+import { Button } from '@/components/ui/button';
 import Input from '@/componenets/Input/Input';
-import Picker from 'react-mobile-picker';
-import {
-  Drawer,
-  DrawerContent,
-  DrawerHeader,
-} from '@/components/ui/drawer';
+import { DatePickerDrawer } from '@/components/ui/datepicker';
 
 export const MilitaryDay: React.FC = () => {
   const [enlistmentDate, setEnlistmentDate] = useState<Date | null>(null);
@@ -17,17 +12,6 @@ export const MilitaryDay: React.FC = () => {
   const [isDrawerOpen, setIsDrawerOpen] = useState(false);
 
   const navigate = useNavigate();
-
-  const today = new Date();
-  const years = Array.from({ length: 100 }, (_, i) => String(today.getFullYear() - i));
-  const months = Array.from({ length: 12 }, (_, i) => String(i + 1));
-  const days = Array.from({ length: 31 }, (_, i) => String(i + 1));
-
-  const [pickerValue, setPickerValue] = useState({
-    year: String(today.getFullYear()),
-    month: String(today.getMonth() + 1),
-    day: String(today.getDate()),
-  });
 
   const formatDate = (date: Date) => {
     const year = date.getFullYear();
@@ -47,16 +31,23 @@ export const MilitaryDay: React.FC = () => {
     setIsDrawerOpen(true);
   };
 
-  const handleConfirm = () => {
-    const { year, month, day } = pickerValue;
-    const date = new Date(Number(year), Number(month) - 1, Number(day));
+  const handleDateConfirm = (date: Date) => {
     if (activePicker === 'enlistment') {
       setEnlistmentDate(date);
     } else if (activePicker === 'discharge') {
       setDischargeDate(date);
     }
-    setIsDrawerOpen(false);
     setActivePicker(null);
+  };
+
+  const getInitialDate = () => {
+    if (activePicker === 'enlistment' && enlistmentDate) {
+      return enlistmentDate;
+    }
+    if (activePicker === 'discharge' && dischargeDate) {
+      return dischargeDate;
+    }
+    return undefined;
   };
 
   return (
@@ -109,57 +100,12 @@ export const MilitaryDay: React.FC = () => {
         />
       </div>
 
-      <Drawer open={isDrawerOpen} onOpenChange={setIsDrawerOpen}>
-        <DrawerContent >
-          <div className="flex flex-col items-center w-full">
-            <div className="w-full h-[200px] flex items-center justify-center">
-              <Picker value={pickerValue} onChange={setPickerValue} height={180} itemHeight={44}>
-                <Picker.Column name="year">
-                  {years.map(year => (
-                    <Picker.Item key={year} value={year}>
-                      {({ selected }) => (
-                        <div className={`w-24 text-center ${selected ? 'text-black font-semibold text-lg' : 'text-gray-400'}`}>
-                          {year}
-                        </div>
-                      )}
-                    </Picker.Item>
-                  ))}
-                </Picker.Column>
-                <Picker.Column name="month">
-                  {months.map(month => (
-                    <Picker.Item key={month} value={month}>
-                      {({ selected }) => (
-                        <div className={`w-24 text-center ${selected ? 'text-black font-semibold text-lg' : 'text-gray-400'}`}>
-                          {month}
-                        </div>
-                      )}
-                    </Picker.Item>
-                  ))}
-                </Picker.Column>
-                <Picker.Column name="day">
-                  {days.map(day => (
-                    <Picker.Item key={day} value={day}>
-                      {({ selected }) => (
-                        <div className={`w-24 text-center ${selected ? 'text-black font-semibold text-lg' : 'text-gray-400'}`}>
-                          {day}
-                        </div>
-                      )}
-                    </Picker.Item>
-                  ))}
-                </Picker.Column>
-              </Picker>
-            </div>
-          </div>
-          <DrawerHeader className="p-4">
-            <Button
-              text="확인"
-              variant="active"
-              onClick={handleConfirm}
-              className="w-full"
-            />
-          </DrawerHeader>
-        </DrawerContent>
-      </Drawer>
+      <DatePickerDrawer
+        isOpen={isDrawerOpen}
+        onOpenChange={setIsDrawerOpen}
+        onConfirm={handleDateConfirm}
+        initialDate={getInitialDate()}
+      />
     </div>
   );
 };
